Don't reinstall addons if we've already installed the latest version.
Added version tracking in an 'installed versions' file. Just add
an 'Installed Versions File' entry to config.ini and the file will
be created and populated the next time WoWAddonUpdater runs.

Then on subsequent update runs we can just read the version number
from the Curse page and if that's the version we've installed there's
no need to thrash Curse's download server.