Added error checks on page parser
I'm not sure what curse has changed, but I was getting -1 returns on find that was putting very large 2,500 character versions including equal signs and line returns. So I added a check for when find returns -1. put a max character limit on versions of 300 characters and replaced line returns and equal signs in versions with less harmful characters.