Fix ElvUI version fetching and change it to use the master branch
A recent GitLab update broke the regex search. After fixing it I discovered that while the master branch is used to update ElvUI, the hash used for checking for a new release is fetched from the development branch, causing the script to download and reinstall what is probably the currently installed version almost every time the script is run.